% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CVP_exp}
\alias{CVP_exp}
\title{CVP Exports}
\format{
A matrix with months (1-12) as rows and years (1980-1999) as columns
}
\usage{
CVP_exp
}
\description{
Total exports for CVP in cms. Value is obtained using CALSIM variable DEL_CVP_EXP.
}
\details{
The flows here are obtained from a CALSIM run designed for the structured
decision making in the delta. The name of the scenario is: Full Obs No Notch Baseline.
}
\keyword{datasets}
