% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_flows}
\alias{delta_flows}
\title{Delta Flows and Diversions}
\format{dataframe with 972 rows and 11 variables:
\describe{
\item{date}{CALSIM II date}
\item{north_delta_inflow_cfs}{C400 + C157, north delta inflow in cubic fete per second}
\item{south_delta_inflow_cfs}{C401B + C504 + C508 + C644, south delta inflow in cubic fete per second}
\item{north_delta_inflow_cms}{north delta inflow in cubic meters per second}
\item{south_delta_inflow_cms}{south delta inflow in cubic meters per second}
\item{north_delta_div_cfs}{D403A + D403B + D403C + D403D + D404, north delta diversions in cubic feet per second}
\item{south_delta_div_cfs}{D418 + D419 + D412 + D410 + D413 + D409B + D416 + D408_OR + D408_VC, south delta diversions in cubic feet per second}
\item{north_delta_div_cms}{north delta diversions in cubic meters per second}
\item{south_delta_div_cms}{south delta diversions in cubic meters per second}
\item{north_delta_prop_div}{north delta inflow / north delta diversions}
\item{south_delta_prop_div}{south delta inflow / south delta diversions}
}}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
delta_flows
}
\description{
A dataset containing the inflow, total diversions, and proportion diverted for the North and South Deltas.

The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River
below Vernalis.
}
\details{
The inflow, diversions, and proportions diverted in the North and South Deltas are represented using
'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.
The nodes and calculation for each watershed are outlined above.


The CALSIM II run is a Reclamation product used to replicate current operations for comparison
with proposed adjustments under an ongoing Endangered Species Act consultation with the National Marine
Fisheries Service. A current NMFS Biological Opinion concluded that, as proposed, CVP and SWP operations
were likely to jeopardize the continued existence of four federally- listed anadromous fish species:
Sacramento River winter-run Chinook salmon, Central Valley spring-run Chinook salmon, California Central Valley
steelhead, and the Southern distinct population segment of the North American green sturgeon. This CALSIM II
run was used as the basis of comparison for other potential operations that could offset impacts to listed species.
More information on CALSIM II is available \href{http://baydeltaoffice.water.ca.gov/modeling/hydrology/CalSim/index.cfm}{here}

\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
