% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_routing_flows}
\alias{delta_routing_flows}
\alias{Q_free}
\alias{Q_vern}
\alias{Q_stck}
\alias{CVP_exp}
\alias{SWP_exp}
\title{Delta Routing Flows}
\format{
A matrix with
\describe{
\item{Q_free}{flow at Freeport (C400) in cms}
\item{Q_vern}{flow at Vernalis (C639) in cms}
\item{Q_stck}{flow at Stockton (C417A) in cms}
\item{CVP_exp}{CVP Exports in cms, uses CALSIM variable DEL_CVP_TOTAL}
\item{SWP_exp}{SWP Exports in cms, uses CALSIM variable DEL_SWP_TOTAL}
}

An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.
}
\usage{
Q_free

Q_vern

Q_stck

CVP_exp

SWP_exp
}
\description{
Flows is cms for the
}
\details{
The flows here are obtained from a CALSIM run designed for the structured
decision making in the delta. The name of the scenario is: Full Obs No Notch Baseline.
}
\keyword{datasets}
