% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeportQ}
\alias{freeportQ}
\title{The Flow from Lower Sacramento River into the Central/South Delta}
\format{dataframe with 996 rows and 3 variables:
\describe{
\item{date}{CALSIM II date}
\item{freeportQcfs}{C400 - flow in cubic feet per second}
\item{freeportQcms}{C400 - flow in cubic meters per second}
}}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
freeportQ
}
\description{
A dataset containing the flow at Freeport Weir. To be used for routing fish from the Lower
Sacramento River into the Central/South delta in the SIT Salmon Lifecycle Model.
}
\details{
The flow at Freeport Weir is represented using node CALSIM II 'FLOW-CHANNEL' C400 and the

The CALSIM II run is a Reclamation product used to replicate current operations for comparison
with proposed adjustments under an ongoing Endangered Species Act consultation with the National Marine
Fisheries Service. A current NMFS Biological Opinion concluded that, as proposed, CVP and SWP operations
were likely to jeopardize the continued existence of four federally- listed anadromous fish species:
Sacramento River winter-run Chinook salmon, Central Valley spring-run Chinook salmon, California Central Valley
steelhead, and the Southern distinct population segment of the North American green sturgeon. This CALSIM II
run was used as the basis of comparison for other potential operations that could offset impacts to listed species.
More information on CALSIM II is available \href{http://baydeltaoffice.water.ca.gov/modeling/hydrology/CalSim/index.cfm}{here}

\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
