% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_pulse_flows}
\alias{prop_pulse_flows}
\title{Proportion Pulse Flow}
\format{
a 31 by 12 matrix (watersheds by months)
}
\usage{
prop_pulse_flows
}
\description{
Estimated monthly proportion of flow that is a pulse
}
\details{
prop_pulse_flows = sd(flow)/median(flow)/100

Calculated using \code{\link{flows_cfs}}
}
\keyword{datasets}
