% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{propQbypass}
\alias{propQbypass}
\title{Proportion of Sacramento River Flow through the Sutter and Yolo Bypasses}
\format{dataframe with 984 rows and 9 variables:
\describe{
\item{date}{CALSIM II date}
\item{propQyolo}{D160/C134}
\item{propQyolo1}{D160/C134}
\item{propQyolo2}{D166A/C165}
\item{propQsutter}{(D117 + D124 + D125 + D126)/C116}
\item{propQsutter1}{D117/C116}
\item{propQsutter2}{D124/C123}
\item{propQsutter3}{D125/C124}
\item{propQsutter4}{D126/C125}
}}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
propQbypass
}
\description{
A dataset containing the proportion of Sacramento River flow within the bypasses for
use with the CVPIA Science Integration Team Decision Support Model to apportion fish onto the bypasses.
}
\details{
The proportions of Sacramento River flowing through the bypasses are represented using
'FLOW-CHANNEL' and 'FLOW-DELIVERY' nodes from CALSIM II.
The nodes and calculation for each watershed are outlined above.

\strong{Model Usage:}

Option 1 - propQyolo and propQsutter represent the proportion of flow as a single value for the bypasses

Option 2 - use the numbered propQ columns to represent the proportion of flow at each weir

The CALSIM II run is a Reclamation product used to replicate current operations for comparison
with proposed adjustments under an ongoing Endangered Species Act consultation with the National Marine
Fisheries Service. A current NMFS Biological Opinion concluded that, as proposed, CVP and SWP operations
were likely to jeopardize the continued existence of four federally- listed anadromous fish species:
Sacramento River winter-run Chinook salmon, Central Valley spring-run Chinook salmon, California Central Valley
steelhead, and the Southern distinct population segment of the North American green sturgeon. This CALSIM II
run was used as the basis of comparison for other potential operations that could offset impacts to listed species.
More information on CALSIM II is available \href{http://baydeltaoffice.water.ca.gov/modeling/hydrology/CalSim/index.cfm}{here}

\href{}{CALSIM II schematic}
}
\keyword{datasets}
