% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return_flow}
\alias{return_flow}
\title{Return Flow}
\format{dataframe with 31 rows and 83 variables:
\describe{
\item{watershed}{31 CVPIA watersheds}
\item{date}{CALSIM II date}
\item{retQ}{proportion of flow at tributary junction from natal watershed in October}
}}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
return_flow
}
\description{
A dataset containing the proportion flows at tributary junction
coming from natal watershed using October CALSIM II flows. These proportions are used to estimate
straying in the CVPIA SIT Salmon Population Model.
}
\details{
The return flow proportion is calculated using the average October flow in each watershed divided by the
average October flow of the tributary it flows into for each year of the simulation. These tributary relationships
are described in detail below.


\strong{Flow into Upper-mid Sacramento River:}
\itemize{
  \item Upper Sacramento River
  \item Antelope Creek
  \item Battle Creek
  \item Bear Creek
  \item Big Chico Creek
  \item Butte Creek
  \item Clear Creek
  \item Cottonwood Creek
  \item Cow Creek
  \item Deer Creek
  \item Elder Creek
  \item Mill Creek
  \item Paynes Creek
  \item Stony Creek
  \item Thomes Creek
}

\strong{Flow into Feather River:}
\itemize{
  \item Bear River
  \item Yuba River
}

\strong{Flow into Lower-mid Sacramento River:}
\itemize{
  \item Feather River
}

\strong{Flow into Lower Sacramento River:}
\itemize{
  \item American River
}

\strong{Flow into San Joaquin River:}
\itemize{
  \item Merced River
  \item Tuolumne River
  \item Stanislaus River
}

The Mokulumne River, Calaveras River, and Cosumnes River are assigened 100\%.


\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
