% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upsacQ}
\alias{upsacQ}
\title{Upper Sacramento River Flow}
\format{
dataframe with 984 rows and 3 variables:
\describe{
\item{date}{CALSIM II date}
\item{upsacQcfs}{C109, flow in cfs}
\item{upsacQcms}{C109, flow in cms}
}
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
upsacQ
}
\description{
A dataset containing the Upper Sacramento River flow in cfs and cms
}
\details{
The Upper Sacramento River is represented using node CALSIM II 'FLOW-CHANNEL' C109 node at Bend.


\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
