% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeportQ}
\alias{freeportQ}
\title{The Flow from Lower Sacramento River into the Central/South Delta}
\format{
dataframe with 996 rows and 3 variables:
\describe{
\item{date}{CALSIM II date}
\item{freeportQcfs}{C400 - flow in cubic feet per second}
\item{freeportQcms}{C400 - flow in cubic meters per second}
}
}
\source{
\itemize{
  \item \strong{Data Wrangling:} Sadie Gill  \email{sgill@flowwest.com}
  \item \strong{Node Selection:} Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
  \item \strong{CALSIM Model Output:} Michael Wright \email{mwright@usbr.gov}
}
}
\usage{
freeportQ
}
\description{
A dataset containing the flow at Freeport Weir. To be used for routing fish from the Lower
Sacramento River into the Central/South delta in the SIT Salmon Population Model.
}
\details{
The flow at Freeport Weir is represented using node CALSIM II 'FLOW-CHANNEL' C400.


\href{https://s3-us-west-2.amazonaws.com/cvpiaflow-r-package/BST_CALSIMII_schematic_040110.jpg}{CALSIM II schematic}
}
\keyword{datasets}
